Update tests to use chai/mocha part 1
