+++ ember-cli-build.js
@@ -4,7 +4,9 @@ var EmberAddon = require('ember-cli/lib/broccoli/ember-addon');
 
 module.exports = function(defaults) {
   var app = new EmberAddon(defaults, {
-    // Add options here
+    'ember-cli-babel': {
+      includePolyfill: true
+    }
   });
 
   /*
