+++ tests/.eslintrc.js
@@ -1,5 +1,10 @@
 module.exports = {
   env: {
     embertest: true
+  },
+  globals: {
+    'server': true,
+    'beforeEach': true,
+    'afterEach': true
   }
-}
+};
