+++ tests/helpers/start-app.js
@@ -2,22 +2,14 @@ import Ember from 'ember'
 import Application from '../../app'
 import config from '../../config/environment'
 
-const {
-  merge,
-  run
-} = Ember
-
 export default function startApp(attrs) {
-  let application
+  let attributes = Ember.merge({}, config.APP)
+  attributes = Ember.merge(attributes, attrs) // use defaults, but you can override;
 
-  let attributes = merge({}, config.APP)
-  attributes = merge(attributes, attrs) // use defaults, but you can override
-
-  run(() => {
-    application = Application.create(attributes)
+  return Ember.run(() => {
+    const application = Application.create(attributes)
     application.setupForTesting()
     application.injectTestHelpers()
+    return application
   })
-
-  return application
 }
