+++ tests/test-helper.js
@@ -1,6 +1,4 @@
 import resolver from './helpers/resolver'
-import {
-  setResolver
-} from 'ember-qunit'
+import { setResolver } from 'ember-mocha'
 
 setResolver(resolver)
